requirements.txt
flask
torch
torchvision
pillow
numpy
pandas
scikit-learn
